# docs/swagger.yaml
definitions:
  model.DurationStatsRequest:
    properties:
      duration:
        description: duration-查询的时间范围 枚举类型：0.昨天 1.过去7天 2.过去30天 3.过去90天 4.上周汇总 5.上月汇总"
          enum:"0,1,2,3,4,5
        maximum: 5
        minimum: 0
        type: integer
      "n":
        description: 检索数量前n的intent信息，n允许范围[1,1000]
        maximum: 1000
        minimum: 1
        type: integer
    required:
    - "n"
    type: object
  model.ExactStatsRequest:
    properties:
      end_time:
        description: 检索范围的结束时间 unix时间戳
        minimum: 0
        type: integer
      "n":
        description: 检索数量前n的intent信息，n允许范围[1,1000]
        maximum: 1000
        minimum: 1
        type: integer
      start_time:
        description: 检索范围的起始时间 unix时间戳
        minimum: 0
        type: integer
    required:
    - end_time
    - "n"
    - start_time
    type: object
  model.IntentResult:
    properties:
      count:
        description: 按当前条件统计到的数量
        type: integer
      intent_id:
        description: 知识点唯一id
        type: string
      intent_name:
        description: 知识点最新名称
        type: string
    type: object
info:
  contact: {}
paths:
  /chatbot/v1alpha1/agents/{agentId}/channels/{channelId}/getReply:
    post:
      consumes:
      - application/json
      description: 入参出参均与原始文档保持一致 https://demo.laiye.com:8083/chatbot-openapi/swagger-ui/#/ChannelReplyService/ChannelReplyService_GetReply
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: body
      summary: 代理机器人Api
      tags:
      - botAgent
  /count:
    get:
      consumes:
      - application/json
      description: 返回bot_response总行数
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: body
      summary: 总记录数
      tags:
      - 开发测试
  /flush:
    get:
      consumes:
      - application/json
      description: 刷新昨日、近7天、近30天、近90天的TopN缓存结果
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: body
      summary: 手动刷新统计数据缓存
      tags:
      - 开发测试
  /stats:
    post:
      consumes:
      - application/json
      description: 点击200 Successful Response查看具体接口返回格式
      parameters:
      - description: 控制参数
        in: body
        name: "n"
        required: true
        schema:
          $ref: '#/definitions/model.DurationStatsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.IntentResult'
            type: array
      summary: 按枚举范围统计知识点TopN
      tags:
      - botAgent
  /topn:
    post:
      consumes:
      - application/json
      description: 点击200 Successful Response查看具体接口返回格式
      parameters:
      - description: ' '
        in: body
        name: "n"
        required: true
        schema:
          $ref: '#/definitions/model.ExactStatsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.IntentResult'
            type: array
      summary: 按精确范围统计知识点TopN
      tags:
      - 开发测试
swagger: "2.0"